patient Complaint

+++ src/app/mainapp/userlist/userlist.component.ts
@@ -3,6 +3,8 @@ import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http
 import { BehaviorSubject, Observable, timer, interval, Subscription } from 'rxjs';
 import { environment } from 'src/environments/environment';
 import { map, catchError } from 'rxjs/operators';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import Swal from "sweetalert2";
 
 import { User } from 'src/app/models/Users';
 import { UserService } from 'src/app/services/user.service';
@@ -29,7 +31,7 @@ export class UserlistComponent {
     'mobile',
     'firstName',
     'lastName',
-    'isActive',
+    'active',
     'id'];
   user = [];
   search = new FormControl();
@@ -40,11 +42,15 @@ export class UserlistComponent {
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
   @ViewChild(MatSort, { static: true }) sortable: MatSort;
 
-  constructor(private http: HttpClient, private userService: UserService, private dialog: MatDialog) { }
+  constructor(private snakbar: MatSnackBar,private http: HttpClient, private userService: UserService, private dialog: MatDialog) { }
 
   ngOnInit() {
     this.dataSource.paginator = this.paginator;
     this.dataSource.sort = this.sortable;
+    this.loadData()
+  }
+
+  loadData(){  
     this.userService.loadData().subscribe(results => {
       console.log(results)
       this.loadEmptyMsg = true;
@@ -52,24 +58,43 @@ export class UserlistComponent {
       this.dataSource.data = results;
     });
   }
+  ondelete(user: User) {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        console.log("delete: " + user.id);
+        this.userService.delete(user.id).subscribe(params => {
+          console.log('come to the subscriber: ');
+          console.log(params);
+    
+          if (params) {
+            this.loadData()
+    
+            this.snakbar.open('User deleted successfully.', 'Dismise', {
+              duration: 3000,
+              horizontalPosition: 'right',
+              verticalPosition: 'top',
+            });
+          }
+        });
 
-  ondelete(user: any) {
-
-    console.log("delete: " + user.id);
-    //console.log($event.checked);
-
-    //user.IsActive = $event.checked;
-    this.userService.delete(user.id).subscribe(params => {
-      console.log('come to the subscriber: ');
-      console.log(params);
-
-      if (params) {
-        const index: number = this.dataSource.data.indexOf(user);
-        if (index !== -1) {
-          this.dataSource.data.splice(index, 1);
-        }
       }
-    });
+      else {
+        Swal.fire(
+          'Cancelled',
+          'Your User is safe :)',
+          'error'
+        )
+      }
+    })
+   
   }
   onRemoveAll() {
 
@@ -103,22 +128,35 @@ export class UserlistComponent {
     });
     dialogRef.afterClosed().subscribe(result => {
       console.log('The dialog was closed');
-
+      if(result)
+      this.loadData()
+      this.snakbar.open('User created successfully.', 'Dismise', {
+        duration: 3000,
+        horizontalPosition: 'right',
+        verticalPosition: 'top',
+      });
     });
   }
 
-  onEdit(id: number) {
+  onEdit(user: User) {
     console.log("Calling edit answer list");
-    console.log(id);
+    console.log(user);
 
     const dialogRef = this.dialog.open(UserdetailComponent, {
       width: '650px',
-      //data: { id: id, sectionID: this.sectionID }
+      data: { user: user }
     });
 
     dialogRef.afterClosed().subscribe(result => {
       console.log('The dialog was closed');
+      if(result)
+      this.loadData()
+      this.snakbar.open('User Updated successfully.', 'Dismise', {
+        duration: 3000,
+        horizontalPosition: 'right',
+        verticalPosition: 'top',
+      });
     });
   }
-  reload() { }
+   
 }
